Renamed to operation_states
